App work with net and API, all basic functions works

+++ lib/domain/weather/weather_components/weather_hourly.dart
@@ -1,7 +1,7 @@
 import 'package:my_weather_app/domain/constants/constants.dart';
 
 class WeatherHourly {
-  List<double> temperature = List(kAmoutHourlyComponents);
+  List<int> temperature = List(kAmoutHourlyComponents);
   List<String> weatherIcon = List(kAmoutHourlyComponents);
 
   // List<double> get temperature => _temperature;
