+++ lib/domain/weather/weather_components/wind.dart
@@ -1,6 +1,6 @@
 class Wind {
   String direction = '';
-  double speed = 0;
+  int speed = 0;
 
   // String get direction => _direction;
   // double get speed => _speed;
